move #!set handling to custom parser, add support for --byref
